0.5.3
General bug fixing. I had to delete some includes because we were
getting multiple includes. Because of that I had to change some of the
functions in Player that relied on access to the Game class which they
no longer have. Those variables are now provided through parameters of
the function. According to the debug the equality operator doesn't work
so I might look into that next.