0.5.4
More Bug Fixing. I think I got everything the compiler is talking about.
The only thing left according to the compiler errors is to write a play
game function. I also added a boolean playsFirst to both Computer and
Player and added constructors that initialize it to false, as well as
set score = 0.